refactor 1、将editor和renderer setting作为application的成员变量保存，仅传入它们的引用； 2、在editor中加入callback以及fonts； 3、按titlebar移动仍需debug。

+++ include/displayinfo.hpp
@@ -3,9 +3,9 @@
 namespace tinyglrenderer {
 
 struct DisplayInfo {
-    float fps         = 0;
-    float deltaTime   = 0;
-    uint32_t drawCall = 0;
+    float fps       = 0;
+    float deltaTime = 0;
+    size_t drawCall = 0;
 };
 
 } // namespace tinyglrenderer
